shiny/driver/gldriver: texture drawing
Adapted from x/mobile/exp/gl/glutil. As shriny.Drawer takes a affine
transform, I've semi-reverted to an earlier version of the glimage
logic that was based on affine transforms. I believe this is correct
(and it can draw the basics on the screen), but it needs more robust
experimentation and testing.

Change-Id: Icc77de0343fe7965b21d6cb4923f521986e088af