Update patch set 1
Patch Set 1:

I'm not blocked on this, it's just an issue I've been thinking about for a while, and I would like to solve it before screen becomes the foundation of x/mobile/app.

 > Do you know how Cocoa OpenGL programs deal with multiple windows
 > and multiple GL contexts? How does something like SDL cope with
 > multiple windows and multiple GL contexts? Or do they not provide
 > that?
 
Given an OpenGL context value, you pin the current OS thread to it by making it current. All use of a context must be single-threaded, but multiple threads can use different contexts concurrently.

If you recall, an earlier design of the gl package had a context object and moved all the functions to be methods on it. This was to replicate this model in Go.

 > One thing we could do, though, is have gldriver's screenImpl.NewWindow
 > return an error after the first one. It should be possible to run
 > shiny on mobile, and on mobile you don't get to have more than one
 > window, so I don't think I wrote it down anywhere, but my intention
 > was always for NewWindow to possibly fail on the second call, on
 > some platforms.
 > 
 > It may be that we provide multiple drivers: if you're writing a
 > game and want a single window or full-screen window that gives you
 > OpenGL, then use gldriver. If you're writing a traditional word
 > processor and want multiple windows but don't need OpenGL, then use
 > x11driver or cocoadriver (backed by something like Core Graphics
 > (??) instead of OpenGL per se?) or win32driver.

We could do that (I've been considering a cocoadriver just to help pull apart and clean up some of the cocoa logic), but it's not obviously better than the code in this CL.

Long term, I think we are limiting ourselves by keeping the GL package to only one active context per process. On the other hand, there almost always is only one (some games use background contexts stream load textures, etc, but not many), and I don't want to complicate the gl package more than is absolutely necessary.

I'm once again leaning towards introducing a gl.Context object, and having x/mobile/app provide the active instance of it, and screen.Window doing the same.

Patch-set: 1