shiny/driver/windriver: new package providing a Windows driver
This provides the initial implementation of a Windows driver for shiny.
It will only open a window and wait for you to close it; Buffer and
Texture will come next, when I figure out how they will work.

I tried to lay down the design of the package in doc.go. If you are still
confused, I'll be glad to rewrite or expand it.

Currently this uses cgo.

Change-Id: I2b060455243f445dd0f4c62f6f0c346768491547