Update patch set 3
Patch Set 3:

I could definitely do all this in pure Go as well. One thing I'll try to do is write this in such a way that migrating to pure Go would be trivial. I would just need four things:

1. A stable, supported DLL calling package (syscall is deprecated, golang.org/x/sys doesn't have one it seems) with documented integer signedness mechanics (so, for instance, CW_USEDEFAULT, defined as ((int)0x80000000), comes out the other end as CW_USEDEFAULT and not some large value on both 32-bit and 64-bit)
2. A guarantee that a pointer to Go data passed into such a syscall will stay alive for the length of the call, regardless of if it's to a variable or to the backing array of a slice
3. A guarantee that the error returned from such a DLL call is the result of GetLastError()
4. Something like syscall.NewCallback() with defined parameter size and signedness semantics (what necessitated me to move to C in the first place for my own package ui was that the one in the deprecated package syscall can't handle more than 4 parameters, and I needed a callback with 6 for subclassing, which is not needed for shiny)

One thing that would help with the C code is a way for the xx, err := C.xxx() syntax to have err be GetLastError() but :/

What you have there with your structure is an interesting way to do things. I was doing something similar to mksyscall earlier myself.

Patch-set: 3