shiny/driver/windriver: implement Window.Fill.
When Windows sends a WM_PAINT message, we enter a loop
where a private channel is used to forward painting requests
to be performed by the main thread. This ensures we don't get
overlapping WM_PAINT messages. Calling EndPaint will end
this loop and allow the program to continue to work.

Hopefully the code is somewhat self-explanatory. The actual
GDI muckery may appear foreign, but that's GDI for you :/

This will be the last commit that introduces new C code.
Past experience has shown that working with GDI exposes
edge cases in the Go syscall implementation, so I wanted
to get at least Window.Fill working in C so I have a known
good point to work off of. If someone doesn't beat me to it,
I'll start converting the code to pure Go tomorrow.

This introduces the condition that all drawing must be done in
response to a paint.Event. If this is not desired, it can be changed
after the C -> Go conversion.

Change-Id: Ic02669d5f7281fac3c11951ce6a590f9d23dd0a1