Update patch set 2
Patch Set 2:

You're probably right; there was a lot of changes that were made as I was writing this, and a few things (like that panic) were remnants of older revisions. I'm not sure if you want me to abandon the CL (and move right to the C rewrite) or not, but I'll respond to a few of the comments anyway.

While a DC can be accessed from multiple threads, DCs have state (SelectObject), and you'll have to synchronize that state if you want to use a DC from multiple threads /at the same time/, so it may be better to lock them after all. The paintRequests channel was ensuring that all drawing operations happen on the main thread, so no locking of w.dc is necessary in the code I gave. I can look into your suggestion for an alternative paint handler, though.

(The only real requirements are that BeginPaint and EndPaint the Windows API functions are called on the main thread and that the DC is not accessed after a call to EndPaint.)

I wasn't aware of the asynchronous nature of the painting commands. I suppose the code as it stands will handle that for us though (possibly even with a buffered channel).

I wasn't aware a nonclosed channel would be garbage collected; that's good to know.

And yes, I could use FillRect instead of the BitBlt for the draw.Src case. I can make that change too.

Patch-set: 2