Update patch set 2
Patch Set 2:

I don't think we need to abandon this CL. It seems fine to land this one first and then convert to Go.

We don't have to implement it in this CL, but I can't remember if we discussed how this paintRequest design will work with multiple windows. If DCs only need to be accessed from a thread and not necessarily the main thread, then I still think the best design is to have a dedicated goroutine, separate from the Windows message loop, that can select on two channels: communication from the OS and from the app. The first one brings the equivalent of WM_PAINT messages (and its DC). The second one brings calls on the Go screen API, like Window.Fill and Window.Upload.

I also can't remember if we discussed how you intend to implement Texture: whether to use GDI+, Direct2D or something else. But I suppose that that can be a separate conversation and code review.

Win32's FillRect (or FillSolidRect??) is probably the way to go, but I did some more nosing around, and read that for AlphaBlend [0], "the source bitmap is stretched to match the destination rectangle", so with the current approach, you'd only need a 1x1 bitmap instead of an MxN one. In any case, though, FillRect is probably best, if you can get it to work.

I also came across "How to: Draw with Opaque and Semitransparent Brushes" [1]. It's .NET, so I don't know exactly how it translates to the lower level Win32 API, but it seems like you can make brushes with alpha. One thing to note is that it looks like Windows uses non-alpha-premultiplied color, since it says "Color.FromArgb(128, 0, 0, 255)", but Go's color.Color interface returns alpha-premultiplied color.

[0] https://msdn.microsoft.com/en-us/library/windows/desktop/dd183351(v=vs.85).aspx

[1] https://msdn.microsoft.com/en-us/library/5s2dwfx1(v=vs.110).aspx

Patch-set: 2