Update patch set 2
Patch Set 2:

> .... If DCs only need to be accessed from a thread and not
 > necessarily the main thread, then I still think the best design is
 > to have a dedicated goroutine, separate from the Windows message
 > loop, that can select on two channels: communication from the OS
 > and from the app. The first one brings the equivalent of WM_PAINT
 > messages (and its DC). The second one brings calls on the Go screen
 > API, like Window.Fill and Window.Upload.

As far as I remember all GUI Windows API needs to happen on the thread that creates window. In our current implementation, that is mainMessagePump function that sits blocked in getMessage. How do you propose to communicate with that thread. It cannot be blocked reading on a channel.

Alex

Patch-set: 2