Update patch set 2
Patch Set 2:

I suspect that we mean the same things. I'm asking if we can call RegisterWindowMessage to get a 'message type' (sorry, not a native Win32 programmer) to use as the "UINT Msg" parameter for SendMessage. As you say, you can then communicate with 'the main thread' from anywhere else, including implementing the Fill method on Window-the-Go-interface.

Or maybe we can just use (WM_USER + x), but I don't know the subtleties or what exactly a "window class" is in terms of Win32.

Patch-set: 2