Update patch set 2
Patch Set 2:

And the utility window already uses WM_USER for this purpose.

For the record, the first thing I'm going to do after this CL is rewrite the utility window in Go, and rename it to the 'Screen window', since that's what its purpose is: providing the functionality of screen.Screen.

Patch-set: 2