Update patch set 2
Patch Set 2:

I'm not sure if we do want Fill to work immediately.

The model I have in mind is that calling Window.Fill for an off-screen window is a no-op. If the window subsequently becomes on-screen (e.g. un-minimized), then we'd get another paint.Event, and in response to that, call Window.Fill again.

Patch-set: 2