Update patch set 2
Patch Set 2:

Yeah, I'll probably get rid of it in a separate change.

So just to make sure (I'll make the changes and update the CL tomorrow), do you want me to get rid of the done channel and queue pending uploads for when a WM_PAINT occurs? I'm not sure how to do that with the code as it is, because I'm sending a paint.Event on WM_PAINT; should I synthesize one at the start of window lifetime like how x11driver does it?

Patch-set: 2