Update patch set 2
Patch Set 2:

Ah, good catch. The easiest thing to do for now is to just do the send without a select, and panic if w.paintRequests is nil.

As I said, there's still an overall shiny TODO to figure out the exact semantics of a common painting model, but we don't have to wait for that to happen before landing this change.

Patch-set: 2