Update patch set 2
Patch Set 2:

Alex: But how do we "just paint it" when Windows technically only wants us to paint in response to a WM_PAINT? We *could* do asynchronous drawing like Chrome does, or we could use a backbuffer for the window. There are many options, but they would require figuring out how paint.Event will effectively work. (Regardless of what we choose, the code in windraw.c (which will be drawing.go after the rewrite) will be the same.)

Patch-set: 2