Update patch set 3
Patch Set 3:

> 1) When do I issue a paint.Event? Nigel already said not to imitate
> x11driver's fake paint.Event; Expose events are similar to WM_PAINT

To be clear, the proper, long term solution, is to convert expose events from the OS to shiny paint.Events, and not to fake a paint.Event.

However, if you just want to make progress, and avoid having to send out a monolithic code change, it's fine to synthesize a one-off paint.Event, as x11driver does.


> 2) Do I need to save the data drawn somewhere else in the case of
> another window overlapping us?

No. When that window no longer overlaps, you'll get a new WM_PAINT, and can re-paint then.


If calling GetDC will work, and painting isn't actually restricted to being inside a WM_PAINT / BeginPaint / EndPaint enclosure, then that's fine (and that's how X11 works), but I am not familiar enough with Win32 programming to say whether that's feasible. Win32's BeginPaint and EndPaint presumably exist for a reason, but I don't know what that reason is.

Patch-set: 3