Update patch set 3
Patch Set 3:

And, yes, "An application should not call BeginPaint except in response to a WM_PAINT message", but I still don't know what BeginPaint does. For example, is it OK to not call BeginPaint and EndPaint at all in response to a WM_PAINT? Is it legit to call GetDC while inside a BeginPaint/EndPaint block or is 'the paint DC' different from 'the regular DC'? (Standard disclaimer, I may be butchering Win32 terminology).

Patch-set: 3