Update patch set 2
Patch Set 2:

The semitransparent brush thing you linked is for GDI+, which is a separate library that does things in software. I believe .net uses this too. While you can have a brush that's a generic bitmap with pure GDI, I don't think it can be used in an alpha-blended way. As far as I could find, with pure GDI AlphaBlend is your only real option. (Yes, GDI+ makes things easier, but it has speed issues, is a C++ library (though there is a flat API), and would be an extra dependency.)

I'd rather not inline the blend function just yet, in case the other drawing operations use it, since the code will be almost identical anyway.

I'll test the GetDC method next and report back with my results.

Patch-set: 2