Update patch set 3
Patch Set 3:

Okay, a naive use of GetDC() doesn't quite work. When the OS damages the window for whatever reason, for instance when it's resized, we lose whatever content was previously there. I could generate a paint.Event on every resize, but that results in flickering (and drawing issues on the initial resize that I'm sure can be fixed easily).

I'm sure this is completely feasible, but we might need to save the drawn data somewhere else and repaint it in WM_PAINT, or figure out how exposes map to paint.Events so things can be repainted.

Perhaps we should save this change for another CL. This one is already doing quite a bit as it is.

If it isn't clear, I'm not trying to argue with what you all want or shoot your ideas down; I support a Go rewrite and I support seeing if we can draw immediately; and in the latter case, it'll be a partial learning experience for me and for everyone who reads our code, which is always welcome. Figuring out how to do these things is the hard part, but we do things not because they are easy...

Patch-set: 3
Label: Code-Review=0