Update patch set 4
Patch Set 4:

Yes; the only question is how to save the drawing requests. Should we just store everything between the first drawing command issued and a call to EndPaint in a []paintRequests that we run through on each WM_PAINT? (This would eliminate the need for paint.Event entirely...) Or should I use a backing bitmap whose size will need to be adjusted should the window ever grow?

Patch-set: 4