Update patch set 4
Patch Set 4:

The final model is that you don't explicitly paint on resize, since if you get a size.Event then presumably there's a paint.Event coming next. (If Windows doesn't do that, then windriver will have to synthesize it).

For windriver, IIUC, EndPaint will essentially be a no-op. I think x11driver will be the same. gldriver is different because it has a notion of swapping buffers, as well as flushing the OpenGL call queue.

Having WM_PAINT do nothing but send a paint.Event and return to the message loop, and using GetDC everywhere else, seems fine to me.

I'm guessing that you don't even have to call BeginPaint / EndPaint, if DefWindowProc will do that for you, but I'm not sure.

Patch-set: 4