Update patch set 4
Patch Set 4:

Having said that, I had a conversation with crawshaw and r earlier today that maybe all the drivers should double-buffer, like the cocoa gldriver currently does, and you definitely wouldn't (as opposed to merely might not) see anything until EndPaint was called. We didn't come to a conclusion, and this CL doesn't need a moving target, but just bear in mind that the overal paint model isn't set in stone yet.

Patch-set: 4
Label: Code-Review=0