Update patch set 4
Patch Set 4:

And as for the discussion you had; that's fine. The actual work of the drawing will be the same regardless of the approach we take; it's just getting what we draw onto the window that'll differ. There will still be a DC at some point in the chain. Especially after the Go rewrite, changes to the paint.Event model will be easier to make this way.

Patch-set: 4