Update patch set 3
Patch Set 3:

> This appears to have broken the shiny/example/tile program.

Sorry about that.

The problem is that as we don't know where the src2dst affine transform is going to put the texture on the window, I'm calling render.Composite with the entire window as a drawing surface. This does not play well with the draw.Src Op specified in the tile example. The parts of the screen for which there are no src are destroyed.

I'm not sure there's an easy way around this, and if there isn't I'm not sure what the proper behavior of the shiny API should be. Is this the correct behavior for the draw.Src given we have src2dst? Or should we ignore the given op in this X11 driver? I think the former.

It also shows why render.Composite takes a bunch of parameters about the destination rectangle, which is what makes the op useful. Hmm.

Patch-set: 3