shiny/driver/x11driver: honor draw.Op where possible
When drawing onto a square region of a window we can use the clipping
built into the X11 Composite command. As we have a complete map from src
to dst for pixels, it's possible to support draw.Src in this case.

For the case of a general affine transform, because X11 works in the
opposite direction to us (we think in terms of src2dst, it thinks in
terms of dst2src), there's no easy way to support draw.Src without doing
the composition ourselves. So instead always use draw.Over in that case.

Fixes example/tile

Change-Id: Id199393c5dd4755bcec5fc2accb6dfb2077861a2