Update patch set 1
Patch Set 1:

Closing the out channel like this CL proposes means that some pending events, such as a lifecycle.StageDead event, might never be sent. This will be problematic for any code that always expects to receive that event, and do some clean-up. Sure, such code is already susceptible to SIGKILL but it'd be nice if e.g. we could always say that clicking on a window's red "X" button will always deliver a lifecycle.StageDead.

Closing the out channel after all pending events are flushed will leak the pumping goroutine if the app code simply stops receiving on that channel. For example, if the closure passed to driver.Main does:

for e := range w.Events() {
  switch e := e.(type) {
  case key.Event:
    if e.Code == key.CodeEscape {
      return // Nobody reads the events channel after this, even if there are events pending.
    }
  etc
  }
}

There may indeed be a better solution, and it is certainly a bug if the shiny.Window docs do not match the implementations. But I don't think this simple fix is the right fix. The right fix might simply be dropping the "Release closes the... event channel" phrase in the doc comment for shiny.Window.Release, but I haven't given golang/go#12828 a lot of thought yet.

Patch-set: 1