Update patch set 1
Patch Set 1:

I think the problem here is we have two ways of signalling the death of a window: a lifecycle.StageDead message and the Events channel being closed.

The message is nice because it is part of the family of lifecycle messages. The closed channel is nice because without it it is too easy for a user to leave a dangling goroutine.

I think failing to deliver other events is OK, it is just StageDead that's valuable. Let me see if there's a way we can have both.

Patch-set: 1