shiny: replace channel with NextEvent method
Instead of ranging over a channel of events provided by the window, a
program processes events by repeatedly calling NextEvent.

While far more pedestrian an interface than the channel, it is not
significantly harder to use and provides an important benefit: the
window object state knows what current event the calling program is
processing.

If we want, with this design we can add extra methods to shiny.Window
that report relevant state for the event. For example, the size of
the window, which otherwise every program ends up tracking itself.

As a bonus, fixes (in shiny, not x/mobile yet) golang/go#12828.

Change-Id: I26bc47822fa0d92b6d8414aa15ba92f6563de9d6