shiny/driver/gldriver: ANGLE amd64 binary
DO NOT SUBMIT
An alternative to consider: load these onto dl.google.com, and have a
binary that uses them download them if they cannot be found on the
system path, and a GOPATH is set. They can be installed in $GOPATH/pkg
just as gomobile does. This would let us keep a big binary blob out of
git and still support "go get".

For golang/go#9306

Change-Id: I69e262436490d57e5067828dbc758eed62eef837