Update patch set 7
Patch Set 7:

(1 comment)

> How is this package suppose to be used on windows? It seems to me
 > we have golang.org/x/exp/shiny/driver/windriver that does the same
 > thing. Why do we need two?
 
There are trade-offs between windriver and gldriver.

The windriver is simpler. Fewer moving lines of code, no extra DLLs.

The gldriver will probably be faster. (I'm not certain about this, if windriver dives into the DirectX APIs it can be as fast.)

But the most valuable feature of gldriver to me is it allows extracurricular graphics programming with an OpenGL ES context. This means if someone wanted to write a graphics-intensive portable program (for example, a computer game, or a 3D CAD program, or something with complex composition requirements), they can use shiny to provide a consistent view of windowing and events across OS X, Windows, and Linux, and use OpenGL ES for (reasonably) portable GPU access.

The 2D composition system exposed by the screen package is reasonably powerful, but only exposes a tiny fraction of what can be done with a GPU. OpenGL ES is a much larger fraction of the GPU.

If we had some basic widgets in shiny, I would consider using windriver to write an installer for the ANGLE DLLs when a program depends on gldriver.

Patch-set: 7