Update patch set 3
Patch Set 3:

About 10% of program starts on El Capitan (OS X 10.11) have a grey screen and never draw. I'm still trying to track this down. It appears to be a problem in our Cocoa setup logic, as the window appears, all the correct events are sent and the paint events succeed.

But regardless, these fixes help.

Patch-set: 3