shiny/driver/gldriver: fix erroneous double-use of sr when uploading to Window
In (w *windowImpl) Upload, the sr parameter is used to calculate a
sub-image of src when uploading src to a temporary texture, t.
The bounding rectangle of t is then (0,0)-(sr.Dx(),sr.Dy()). The same
sr is then used as the sr parameter to w.Draw when drawing the
temporary texture to the window. However, in the case that sr.Min is
not (0,0), this means that t is incorrectly clipped, and the effect
is that part of the image that was inteded to be uploaded to the
window is missing.

Instead, use t.Bounds() as the sr parameter to w.Draw.

Change-Id: Ib526db0c66856c9ffdd939c943dc9063ffc47c25