shiny/driver/gldriver: fix handling of escape and tab keys
In order to handle the escape and tab keys correctly, key event
handling needs to be moved from an NSResponder implementation
to the NSOpenGLView implementation, and handled with the
performKeyEquivalent method. Without this change, pressing escape
triggers the system beep rather than a keyDown event, and the tab
key yields a keyUp event but no keyDown event.

Change-Id: Ieb28274338249352d08325c0f6eac1d4a4a1c98e