shiny/driver/gldriver: use makeFirstResponder
Set view as the first responder via makeFirstResponder, rather
than setting window.nextResponder to view. This prevents a
segmentation fault when handling mouse events. I am not a cocoa
expert and I don't know why this fixes the problem, but it does
(at least in my tests) and after reviewing the docs it seems
to be what we should have been doing in the first place.

Fixes golang/go#13895.

Change-Id: Idfba7515dd26b66320c42c8e91a104138ec833eb