shiny/screen: introduce WindowParameter
WindowParameter replaces NewWindowOptions with an options interface that
can grow over time and does not suffer from the problematic default
values of NewWindowOptions. It can also be used to get and set the
current state of a Window via methods, without adding methods for every
possible parameter.

For golang/go#13339.

Change-Id: Iae1ce6eec433600410b1d24b0957358fb3f7518c