x/exp/shiny/driver/x11driver: don't try to create zero-size buffers
shmOpen fails on size == 0, and gives a cryptic error message
"x11driver: shmOpen: invalid argument". If zero-size buffers are
invalid, we should report them as invalid. This change fails earlier
on zero-size buffers and also adds additional information for
failures in shmOpen.

This change also updates the documentation of the screen package
to be explicit about this requirement.

Change-Id: I6c2fd7885418c230979cb87a04a62abc5384c20c