Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @nigel you said that calling LockOSThread inside the driver is is dangerous. But it looks like windows version of golang.org/x/exp/shiny/driver.Main calls LockOSThread. Is that a problem too?
> 
> Alex

The idea is that driver.Main is called directly from main and does not return into the program exits. That's not as well documented as it should be.

(It is used by the gldriver on OS X to not only lock an OS thread, but to lock the original thread that starts the Go program, as Cocoa ascribes it special powers.)

If you want to lock an OS thread, you can start a goroutine:

go func() {
    runtime.LockOSThread()
    defer runtime.UnlockOSThread()
    // do the relevant memory work
    close(wait)
}()
<-wait

I'm not sure if that makes sense here. Counter-intuitively, goroutines are lighter weight than many graphical operations, possibly even this operation, but you're in a better place to say.

Otherwise handing this work off to the "window thread" seems reasonable.

Patch-set: 3
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0