Update patch set 1
Patch Set 1:

I have a few questions regarding the API design.

0. should we allow the user to configure MSB/LSB-first for
the word transfers? it's occasionally useful, and is configurable
for certain backends.

1. what about multiple chip selects for the same SPI channel?
I know the spidev Linux driver abstrcats away the CS in the
device node name, but for other backends, e.g. libftdi, this might
not the case.

2. if CSs are included in the API, then we have further problems to
consider:
(a) how to set the polarity of the chip selects?
(b) master/slave mode?
(c) should Do hold CS asserted for the entire transfer or separately
for each word transfer?
(d) should the clock be free-running between assertions of CS?

3. should we add a send-only API so the user can repeatedly send
the same payload to a SPI device. Do will always clobber the
byte slice buffer, so it has be refilled each time.

What backends do you plan to work on? libftdi? bit-banging?
BTW, are you interested in a JTAG package?

Patch-set: 1