Update patch set 1
Patch Set 1:

> I have a few questions regarding the API design.
 > 
 > 0. should we allow the user to configure MSB/LSB-first for
 > the word transfers? it's occasionally useful, and is configurable
 > for certain backends.
 
Yes, it should be a part of the API. SetDataOrder method is what I am planning to add next.

 > 1. what about multiple chip selects for the same SPI channel?
 > I know the spidev Linux driver abstrcats away the CS in the
 > device node name, but for other backends, e.g. libftdi, this might
 > not the case.
 
The current API is a devfs implementation. I will revise the design to depend on a bus and chip select rather than a filename.

 > 2. if CSs are included in the API, then we have further problems to
 > consider:
 > (a) how to set the polarity of the chip selects?

An API probably needed to configure this.

 > (b) master/slave mode?

I don't have a desire to include master management in the scope of this project.
I also haven't thought of writing to the SS pin directly from this library. Could it be why you are asking the question above?

 > (c) should Do hold CS asserted for the entire transfer or
 > separately
 > for each word transfer?

For the entire transfer.

 > (d) should the clock be free-running between assertions of CS?
 
I don't think this question is in the scope of the Go SPI package. Chips allow clock to be running freely between assertions of CS but we never have to care the actual state of the SPI pins. We will depend on higher level drivers to do the work.

 > 3. should we add a send-only API so the user can repeatedly send
 > the same payload to a SPI device. Do will always clobber the
 > byte slice buffer, so it has be refilled each time.
 > 
 > What backends do you plan to work on? libftdi? bit-banging?
 > BTW, are you interested in a JTAG package?

devfs, mraa, a custom Brillo backend and an experimental firmata SPI bridge.

Patch-set: 1