shiny/widget: add Palette and Theme types.
The exact colors used for the default palette is arbitrary, but these
Material Design colors seem as good a choice as any other for now.

A Palette's fields roughly corresponds to what a Java Swing UI looks
like. It might change later to accomodate something more modern, but it
will do for now in order to make progress with upcoming widget code.

Change-Id: Ibd6452e91ae281244adcaf9c91d8fbef023307b9