Update patch set 1
Patch Set 1:

(1 comment)

I think it should be possible to implement most of a back buffer in a separate package, representing that backbuffer as a screen.Texture, instead of being so X11-specific, representing it as an xproto.Pixmap.

The point of doing it as a separate package is that the windriver also needs a back buffer, and I suspect that they can share the (screen.Texture based) implementation.

Patch-set: 1