io/i2c: remove command argument from Read and Write
I2C devices support command-less transmission. This CL is removing the
command argument from Read and Write methods.
Users still can pass it as the first byte in the buf if the device
requires a command.

Also, switching back to read(2) and write(2) calls since the I2C dev
interface document is listing them as an equivalent of the IOCTLs.

Change-Id: Icc236ebac0f92d471ac5f13205fa5f9882fef07d