io/i2c: remove command argument from Read and Write
I2C devices support command-less transmission, hence removeing the
command argument. Users still can pass it as the first byte in the buf
if the device requires a command.

Switching back to old read(2) and write(2) calls since the I2C dev
interface is listing them as an equivalent of the IOCTLs.

Change-Id: Icc236ebac0f92d471ac5f13205fa5f9882fef07d