Update patch set 4
Patch Set 4:

There are alternative ways to handle 10-bit address:
because we take an int which is at least 32-bit
for the address, we can use a special bit marker
for 10-bit addresses. E.g. 0x10000 if set signifies
10-bit addresses. And we provide a constructor
for 10-bit addresses. (I think the Linux kernel uses
this, but it uses 0xa0000 for the marker IIRC.)

The API will look something like this:


  dev, err := i2c.Open(&i2c.Devfs{}, 1, 0x20) // 7-bit address
  dev, err := i2c.Open(&i2c.Devfs{}, 1, i2c.TenBit(0x20)) // 10-bit address


Or we can make Open take an Addr interface. This
is heavier than the previous solution.

What do you think?

Patch-set: 4
Label: TryBot-Result=0