flex: new package for shiny flex layout
A partial implementation of the CSS flexible layout algorithm as a
shiny layout widget.

I have focused on keeping the implementation as close to the
algorithm described in the specification as possible. As the shiny
widget model has not solidified yet, it may be we want to add one of
the missing flex features in the future (such as inline-axis), and
this makes it relatively straightfoward.

This could use a lot more tests. After this CL I intend to spend a
bit of time porting over the more interesting cases from the W3C
test suite.

Change-Id: I9c1af5f9b4b9c6346722b98c48eca521d7c4563d