Update patch set 1
Patch Set 1:

(5 comments)

Just a quick pass for now:

This is doing two separable things: adding new Foo and FooClass widget-related types outside of the widget package, and then implementing a non-trivial chunk of the Flex algorithm.

Can you do the first one as a separate CL, that implements as basic a layout algorithm as possible, that still needs the type structure it does? There's more than enough to review in that.

Patch-set: 1
Label: TryBot-Result=0