Update patch set 1
Patch Set 1:

> Perhaps I misunderstand, but it looks like this is drawing the same
 > set of pixels as before?

It is effectively drawing the same set of pixels as before, but only because the draw operator blends dst and src (as opposed to the render.PictOpSrc operator). Before, each render.Composite call touched every pixel in the window, but for a lot of them you're compositing a completely transparent pixel on top of the dst pixel. That's a no-op, but it's possibly not a free no-op, multiplied by O(width * height) pixels. Maybe the X11 server implementation is smart about it, but maybe it's not. I suspect the way that I'm using X11/Render isn't entirely what the designers expected, based on the PictOpOutReverse+render.PictOpOver dance I have to do, and therefore might not be optimized. You can see the difference in the bounding AABB rectangle, before and after this CL, if you replace render.PictOpOver with render.PictOpSrc in the final line of textureImpl.draw.

It should also simplify the math a little in an upcoming CL, as the origin is the first TriFan point, not the window's top-left.

This also simplifies the args a little, as you no longer need to pass the window width and height.

This is basically the "TODO: tighten the (0, 0)-(w, h) dst rectangle. As it is, we're compositing an unnecessarily large number of pixels" that was at the bottom of textureImpl.draw.

Patch-set: 1
Label: Code-Review=0