shiny/driver/x11driver: set the texture repeat to "pad".
This means that out-of-bound texture reads return the in-bound edge
values instead of returning zero. Since the picture filter is
"bilinear", this is noticeable when scaling the entirety of the texture
(when srcRect = tex.Bounds()), such as in the example/fluid program.
Before this commit, the edges in the destination window were feathered.
After this commit, they are crisp, the same as the gldriver.

See section 9 "Source and Mask Transformations" of
https://www.x.org/releases/current/doc/renderproto/renderproto.txt

Change-Id: I4e4827d35311790bac0cd092fecb3953707976cb