Update patch set 6
Patch Set 6:

> > I don't think we need a helper WriteReg method. Writing to a
 > > register is syntactically pleasant with Write.
 > >
 > > d.Write([]byte{reg, 0x55, 0x44}
 > 
 > I agree that we don't need it but I think it would be nice for
 > developers to have a symmetrical API. Without that, a developer
 > would need to have more knowledge and realize that a normal write
 > would work. What's the downside of adding a symmetrical API besides
 > having an extra exported method?

It is a symmetric API but not a equally performant implementation because every call to WriteReg will allocate. The users will end up using Write eventually. We need a strategy that avoids them to fall back to Write for performance. Otherwise, we can't achieve what proppy is suggesting.

Patch-set: 6
Label: Code-Review=0