Update patch set 1
Patch Set 1:

> (1 comment)

Identifying pins with labels was beyond the scope of this CL and it is not just about mappings. Currently, the pins are identified with integers.

Switching to string identifiers could be an option but it is not a blocker to have int pins. You are not forced to use "AIN0" for AIN0 and can use an arbitrary integer.

From a developer experience perspective, we can move to string identifiers though. It just should be a CL of its own.

Patch-set: 1