Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

wouldn't map[string]{"1": 1} work in such a case. It might not be super pretty but it should work. I guess the alternative would be to have a more complex mapping with each entry having a physical and alias name mapping to a pin number and then double the APIs. Or is there a cleaner solution I am not thinking about?

Patch-set: 1