Update patch set 1
Patch Set 1:

(1 comment)

> I haven't looked at the code, but I think
 > we need more thoughts on the name of the
 > new methods: we are not limited by what
 > the Linux kernel driver calls the feature.

That is a great point. I haven't used another SPI implementation in many years so I don't even know if this is a common feature that might have different names elsewhere.

 > Maybe we can use a .Begin() and .End() method
 > around the Tx() calls instead? But the word
 > "transaction" has already got another meaning
 > here.
 > 
 > I imagine something like this:
 > 
 > dev.Select()
 > defer dev.Deselect()
 > dev.Tx(...)
 > for { dev.Tx(...) }
 > 
 > And we can document that Tx without Select will
 > automatically select and deselect the device
 > (i.e. single shot transactions).

One potential problem I see with this idea is that there isn't (afaict) a way to just select or deselect the SPI device with the Linux driver without also transmitting data in the process. So the Select call above wouldn't do anything except set an in-memory flag to not auto-deselect, and the Deselect call would still have to make a syscall with a full payload (including tx/rx).

I might be wrong on that though. And otherwise, it sounds like a good alternative.

There's one more thing I need to do to make this multi-Tx use case work well. The device I'm communicating with requires a 1ms break between SPI transactions, but that's not true if I'm just calling Tx repeatedly within a single request/response cycle while waiting for the response bytes. I don't need to sleep at all between those AFAIK. So the other thing my code will end up doing is setting the Delay to 0 before the first Tx call, and then set Delay to 1ms just before my final empty Tx call (which is otherwise used just to deselect the device).

Thought that would be worth calling out in case it's also a common thing that should be made easier.

Patch-set: 1
Label: Code-Review=0