decimal: initial commit
This initial CL is to test interest in an exp/decimal subpackage.
(Per Andrew Gerrard's comment golang/go#233797360, although not necessarily an
endorsement.)

In golang/go#12127 a math/big.Decimal type was proposed. I and many others
thought it was a good idea. There are two difficult parts:

- Choosing the proper decimal layout (How many bits? What encoding?)
- Writing the decimal library (Do we follow the math/big API?)

I've written github.com/ericlagergren/decimal and I'm willing to move it
or parts of it (whatever is needed...) to exp/decimal. It's a "big"
decimal, so the only limits on precision are, more or less, those put forth by
your physical machine.

Features include:

- API mirror's math/big
- No string conversions to do arithmetic (unlike other "big" decimal libraries
in Go and other languages)
- Uses hardware routines if it's able, otherwise it uses the software arithmetic
routines from math/big.Int
- Pretty decent tests
- Some extra math routines as well

Change-Id: Ifcb6023e91d86703888f753cb3f9417c8399ab9b