Update patch set 1
Patch Set 1:

> What does it enable?

It enables theme-dependent color styling such as "this button's background is the theme accent color" or "this label's text is the theme foreground color".

Currently, you can pre-calculate colors from a theme, and simply make a widget.Uniform of a certain blue when you construct the widget tree, but that doesn't allow for changing the theme dynamically.

Patch-set: 1
Label: Code-Review=0