shiny/driver/x11driver: fix ipcCreat constant.
That constant value had a typo. It's value (defined in
/usr/include/linux/ipc.h) is in octal, not hexadecimal.

"man shmget" suggests that the IPC_CREAT flag isn't necessary if the key
is IPC_PRIVATE, as has proven in practice by the current code passing
the wrong flag bit all along.

The best fix, for the Go code, is to delete all mentions of ipcCreat.

Change-Id: I1ec701b3069c35445058c1149f2e9d15bbf4cf4c