Update patch set 4
Patch Set 4: Code-Review-1

(17 comments)

I haven't followed this closely enough, but as is, it's hard to say much here. If there's a standard that's followed it should be mentioned in the documentation. The internal representation should be explained in painstaking detail as it will inform the implementation of the operations. Finally, there should be at least one or two basic operations implemented (SetString, Add, Cmp?) so one can see how it all would play out.

Patch-set: 4
Label: Code-Review=-1